+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - `variant()` method for field reader and `Variant` enum for fields with reserved values
 
+- Field readers and writers use one enum where it is possible
+
 ## [v0.15.2] - 2019-07-29
 
 - No changes, just fixing the metadata since crates.io didn't like the keywords